requirements.txt
Flask
WTForms
flask-wtf
Pillow
numpy
requests
Flask-Images
opencv-python